requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
qiskit>=1.0.0
qiskit-aer>=0.12.0
numpy>=1.24.0
websockets>=10.0
jinja2>=3.0.0
python-multipart>=0.0.5
aiofiles>=0.8.0
cryptography>=41.0.0
Pillow>=10.0.0
six>=1.16.0
pyOpenSSL>=23.0.0